+++ glue/dialogs/link_editor/qt/tests/test_link_editor.py
@@ -3,11 +3,10 @@ from mock import patch
 from qtpy import QtWidgets
 from glue.utils.qt import process_events
 from glue.core import Data, DataCollection
-from glue.core.component_link import identity
 from glue.dialogs.link_editor.qt import LinkEditor
 from glue.core.component_link import ComponentLink
 from glue.plugins.coordinate_helpers.link_helpers import Galactic_to_FK5, ICRS_to_Galactic
-from glue.core.link_helpers import functional_link_collection
+from glue.core.link_helpers import identity, functional_link_collection
 
 
 def non_empty_rows_count(layout):
@@ -17,10 +16,10 @@ def non_empty_rows_count(layout):
     count = 0
     for row in range(layout.rowCount()):
         for col in range(layout.columnCount()):
-            if layout.itemAtPosition(row, col) is not None:
-                if layout.itemAtPosition(row, col).widget() is not None:
-                    count += 1
-                    break
+            item = layout.itemAtPosition(row, col)
+            if item is not None and item.widget() is not None and item.widget().isVisible():
+                count += 1
+                break
     return count
 
 
@@ -31,10 +30,6 @@ def get_action(link_widget, text):
                 if action.text() == text:
                     return action
     raise ValueError("Action '{0}' not found".format(text))
-
-
-def get_link_io(widget):
-    return widget.link_io.itemAt(0).widget().layout()
 
 
 class TestLinkEditor:
@@ -66,7 +61,7 @@ class TestLinkEditor:
         link_widget.state.data2 = self.data1
 
         assert link_widget.button_add_link.isEnabled()
-        assert link_widget.button_remove_link.isEnabled()
+        assert not link_widget.button_remove_link.isEnabled()
 
         dialog.accept()
 
@@ -82,7 +77,7 @@ class TestLinkEditor:
         assert link_widget.state.data1 is self.data1
         assert link_widget.state.data2 is self.data2
         assert link_widget.button_add_link.isEnabled()
-        assert link_widget.button_remove_link.isEnabled()
+        assert not link_widget.button_remove_link.isEnabled()
         dialog.accept()
         assert len(self.data_collection.external_links) == 0
 
@@ -105,7 +100,8 @@ class TestLinkEditor:
         # and nothing on the right.
         assert link_widget.listsel_current_link.count() == 0
         assert link_widget.link_details.text() == ''
-        assert link_widget.link_io.itemAt(0) is None
+        assert non_empty_rows_count(link_widget.combos1) == 0
+        assert non_empty_rows_count(link_widget.combos2) == 0
 
         # Let's add an identity link
         add_identity_link.trigger()
@@ -117,17 +113,18 @@ class TestLinkEditor:
         # right hand panel.
         assert link_widget.listsel_current_link.count() == 1
         assert link_widget.link_details.text() == 'Link conceptually identical components'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'a'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'a'
 
         # Let's change the current components for the link
         link_widget.state.current_link.x = self.data1.id['y']
         link_widget.state.current_link.y = self.data2.id['b']
 
         # and make sure the UI gets updated
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'b'
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'y'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'b'
 
         # We now add another link of a different type
         add_lengths_volume_link.trigger()
@@ -138,11 +135,20 @@ class TestLinkEditor:
         # and make sure the UI has updated
         assert link_widget.listsel_current_link.count() == 2
         assert link_widget.link_details.text() == 'Convert between linear measurements and volume'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 7
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(2, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(3, 1).widget().currentText() == 'z'
-        assert get_link_io(link_widget).itemAtPosition(6, 1).widget().currentText() == 'a'
+        assert non_empty_rows_count(link_widget.combos1) == 3
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos1.itemAtPosition(1, 1).widget().currentText() == 'y'
+        assert link_widget.combos1.itemAtPosition(2, 1).widget().currentText() == 'z'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'a'
+
+        # Try swapping the order of the data, the current link should stay the same
+        link_widget.state.flip_data()
+        assert link_widget.link_details.text() == 'Convert between linear measurements and volume'
+
+        # And flip it back
+        link_widget.state.flip_data()
+        assert link_widget.link_details.text() == 'Convert between linear measurements and volume'
 
         # Now switch back to the first link
         link_widget.state.current_link = type(link_widget.state).current_link.get_choices(link_widget.state)[0]
@@ -150,9 +156,10 @@ class TestLinkEditor:
         # and make sure the UI updates and has preserved the correct settings
         assert link_widget.listsel_current_link.count() == 2
         assert link_widget.link_details.text() == 'Link conceptually identical components'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'b'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'y'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'b'
 
         # Next up, we try changing the data
 
@@ -162,7 +169,8 @@ class TestLinkEditor:
 
         assert link_widget.listsel_current_link.count() == 0
         assert link_widget.link_details.text() == ''
-        assert link_widget.link_io.itemAt(0) is None
+        assert non_empty_rows_count(link_widget.combos1) == 0
+        assert non_empty_rows_count(link_widget.combos2) == 0
 
         # Add another identity link
         add_identity_link.trigger()
@@ -173,9 +181,10 @@ class TestLinkEditor:
         # Now there should be one link in the main list
         assert link_widget.listsel_current_link.count() == 1
         assert link_widget.link_details.text() == 'Link conceptually identical components'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'i'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'a'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'i'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'a'
 
         # Switch back to the original data
         link_widget.state.data1 = self.data1
@@ -183,9 +192,10 @@ class TestLinkEditor:
         # And check the output is as before
         assert link_widget.listsel_current_link.count() == 2
         assert link_widget.link_details.text() == 'Link conceptually identical components'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'b'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'y'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'b'
 
         # Let's now remove this link
         link_widget.button_remove_link.click()
@@ -196,11 +206,12 @@ class TestLinkEditor:
         # We should now see the lengths/volume link
         assert link_widget.listsel_current_link.count() == 1
         assert link_widget.link_details.text() == 'Convert between linear measurements and volume'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 7
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(2, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(3, 1).widget().currentText() == 'z'
-        assert get_link_io(link_widget).itemAtPosition(6, 1).widget().currentText() == 'a'
+        assert non_empty_rows_count(link_widget.combos1) == 3
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos1.itemAtPosition(1, 1).widget().currentText() == 'y'
+        assert link_widget.combos1.itemAtPosition(2, 1).widget().currentText() == 'z'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'a'
 
         dialog.accept()
 
@@ -367,26 +378,29 @@ class TestLinkEditor:
 
         assert link_widget.listsel_current_link.count() == 1
         assert link_widget.link_details.text() == ''
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'c'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'c'
 
         link_widget.state.data1 = self.data3
 
         assert link_widget.listsel_current_link.count() == 2
         assert link_widget.link_details.text() == ''
-        assert non_empty_rows_count(get_link_io(link_widget)) == 6
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'a'
-        assert get_link_io(link_widget).itemAtPosition(2, 1).widget().currentText() == 'b'
-        assert get_link_io(link_widget).itemAtPosition(5, 1).widget().currentText() == 'j'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 2
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'j'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'a'
+        assert link_widget.combos2.itemAtPosition(1, 1).widget().currentText() == 'b'
 
         link_widget.state.current_link = type(link_widget.state).current_link.get_choices(link_widget.state)[1]
 
         assert link_widget.listsel_current_link.count() == 2
         assert link_widget.link_details.text() == ''
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'i'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'c'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'i'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'c'
 
         dialog.accept()
 
@@ -430,11 +444,12 @@ class TestLinkEditor:
 
         assert link_widget.listsel_current_link.count() == 1
         assert link_widget.link_details.text() == 'Link ICRS and Galactic coordinates'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 7
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(2, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(5, 1).widget().currentText() == 'a'
-        assert get_link_io(link_widget).itemAtPosition(6, 1).widget().currentText() == 'b'
+        assert non_empty_rows_count(link_widget.combos1) == 2
+        assert non_empty_rows_count(link_widget.combos2) == 2
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos1.itemAtPosition(1, 1).widget().currentText() == 'y'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'a'
+        assert link_widget.combos2.itemAtPosition(1, 1).widget().currentText() == 'b'
 
         dialog.accept()
 
@@ -466,11 +481,12 @@ class TestLinkEditor:
 
         assert link_widget.listsel_current_link.count() == 1
         assert link_widget.link_details.text() == 'Link Galactic and FK5 (J2000) Equatorial coordinates'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 7
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(2, 1).widget().currentText() == 'y'
-        assert get_link_io(link_widget).itemAtPosition(5, 1).widget().currentText() == 'c'
-        assert get_link_io(link_widget).itemAtPosition(6, 1).widget().currentText() == 'b'
+        assert non_empty_rows_count(link_widget.combos1) == 2
+        assert non_empty_rows_count(link_widget.combos2) == 2
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos1.itemAtPosition(1, 1).widget().currentText() == 'y'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'c'
+        assert link_widget.combos2.itemAtPosition(1, 1).widget().currentText() == 'b'
 
         dialog.accept()
 
@@ -503,7 +519,7 @@ class TestLinkEditor:
 
         link_widget.state.current_link.x = self.data1.id['y']
 
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'y'
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'y'
 
         add_identity_link = get_action(link_widget, 'identity')
         add_identity_link.trigger()
@@ -559,17 +575,23 @@ class TestLinkEditor:
 
         assert link_widget.listsel_current_link.count() == 2
 
+        assert not link_widget.combos1_header.isVisible()
+        assert not link_widget.combos2_header.isVisible()
         assert link_widget.link_details.text() == 'Legacy link'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 4
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(2, 1).widget().currentText() == 'c'
+        assert non_empty_rows_count(link_widget.combos1) == 2
+        assert non_empty_rows_count(link_widget.combos2) == 0
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos1.itemAtPosition(1, 1).widget().currentText() == 'c'
 
         link_widget.state.current_link = type(link_widget.state).current_link.get_choices(link_widget.state)[1]
 
+        assert link_widget.combos1_header.isVisible()
+        assert link_widget.combos2_header.isVisible()
         assert link_widget.link_details.text() == 'New-style link'
-        assert non_empty_rows_count(get_link_io(link_widget)) == 5
-        assert get_link_io(link_widget).itemAtPosition(1, 1).widget().currentText() == 'x'
-        assert get_link_io(link_widget).itemAtPosition(4, 1).widget().currentText() == 'c'
+        assert non_empty_rows_count(link_widget.combos1) == 1
+        assert non_empty_rows_count(link_widget.combos2) == 1
+        assert link_widget.combos1.itemAtPosition(0, 1).widget().currentText() == 'x'
+        assert link_widget.combos2.itemAtPosition(0, 1).widget().currentText() == 'c'
 
         dialog.accept()
 
@@ -608,3 +630,20 @@ class TestLinkEditor:
 
         assert link_widget.state.data1 == self.data1
         assert link_widget.state.data2 == self.data2
+
+    def test_preexisting_links_twodata(self):
+
+        # Regression test for an issue that occurred specifically if there were
+        # exactly two datasets and pre-existing links (since this means that
+        # the window opens with a current_link selected by default)
+
+        data1 = Data(x=[1, 2, 3], y=[2, 3, 4], z=[6, 5, 4], label='data1')
+        data2 = Data(a=[2, 3, 4], b=[4, 5, 4], c=[3, 4, 1], label='data2')
+
+        data_collection = DataCollection([data1, data2])
+
+        link1 = ComponentLink([data1.id['x']], data2.id['c'])
+        data_collection.add_link(link1)
+
+        dialog = LinkEditor(data_collection)
+        dialog.show()