+++ glue/dialogs/subset_facet/qt/subset_facet.py
@@ -4,89 +4,97 @@ import os
 import numpy as np
 from matplotlib import cm
 
-from qtpy import QtWidgets, QtGui
+from qtpy import QtWidgets
 from glue.core.util import colorize_subsets, facet_subsets
+from glue.core.state_objects import State
+from glue.external.echo import CallbackProperty, SelectionCallbackProperty
 from glue.utils.qt import load_ui
-from glue.utils.qt.widget_properties import (ButtonProperty, FloatLineProperty,
-                                             ValueProperty)
-from glue.utils.array import pretty_number
-from glue.utils import Pointer, nanmin, nanmax
-from glue.utils.qt import cmap2pixmap
+from glue.core.data_combo_helper import DataCollectionComboHelper, ComponentIDComboHelper
+from glue.external.echo.qt import autoconnect_callbacks_to_qt
+from glue.core.state_objects import StateAttributeLimitsHelper
 
-# We do the following import to register the custom Qt Widget there
-from glue.dialogs.common.qt import component_selector  # noqa
-
-__all__ = ['SubsetFacet']
+__all__ = ['SubsetFacetDialog']
 
 
-class SubsetFacet(QtWidgets.QDialog):
+class SubsetFacetState(State):
 
-    log = ButtonProperty('ui.checkbox_log')
-    vmin = FloatLineProperty('ui.value_min')
-    vmax = FloatLineProperty('ui.value_max')
-    steps = ValueProperty('ui.value_n_subsets')
-    data = Pointer('ui.component_selector.data')
-    component = Pointer('ui.component_selector.component')
+    log = CallbackProperty(False)
+    v_min = CallbackProperty(0.)
+    v_max = CallbackProperty(1.)
+    steps = CallbackProperty(5)
+    data = SelectionCallbackProperty()
+    att = SelectionCallbackProperty()
+    cmap = CallbackProperty()
+
+    def __init__(self, data_collection):
+
+        super(SubsetFacetState, self).__init__()
+
+        self.data_helper = DataCollectionComboHelper(self, 'data', data_collection)
+        self.att_helper = ComponentIDComboHelper(self, 'att')
+        self.lim_helper = StateAttributeLimitsHelper(self, attribute='att',
+                                                     lower='v_min', upper='v_max',
+                                                     log='log')
+
+        self.add_callback('data', self._on_data_change)
+        self._on_data_change()
+
+    def _on_data_change(self, *args, **kwargs):
+        self.att_helper.set_multiple_data([] if self.data is None else [self.data])
+
+
+class SubsetFacetDialog(QtWidgets.QDialog):
+    """
+    Create a new dialog for subset faceting
+
+    Parameters
+    ----------
+    collect : :class:`~glue.core.data_collection.DataCollection`
+        The data collection to use
+    default : :class:`~glue.core.data.Data`, optional
+        The default dataset in the collection (optional)
+    """
 
     def __init__(self, collect, default=None, parent=None):
-        """Create a new dialog for subset faceting
 
-        :param collect: The :class:`~glue.core.data_collection.DataCollection` to use
-        :param default: The default dataset in the collection (optional)
-        """
+        super(SubsetFacetDialog, self).__init__(parent=parent)
 
-        super(SubsetFacet, self).__init__(parent=parent)
+        self.state = SubsetFacetState(collect)
 
         self.ui = load_ui('subset_facet.ui', self,
                           directory=os.path.dirname(__file__))
-        self.ui.setWindowTitle("Subset Facet")
+        self._connections = autoconnect_callbacks_to_qt(self.state, self.ui)
+
         self._collect = collect
 
-        self.ui.component_selector.setup(self._collect)
         if default is not None:
-            self.ui.component_selector.data = default
+            self.state.data = default
 
-        val = QtGui.QDoubleValidator(-1e100, 1e100, 4, None)
-        self.ui.component_selector.component_changed.connect(self._set_limits)
+        self.state.cmap = cm.RdYlBu
 
-        combo = self.ui.color_scale
-        for cmap in [cm.cool, cm.RdYlBu, cm.RdYlGn, cm.RdBu, cm.Purples]:
-            combo.addItem(QtGui.QIcon(cmap2pixmap(cmap)), cmap.name, cmap)
-
-    def _set_limits(self):
-        data = self.ui.component_selector.data
-        cid = self.ui.component_selector.component
-
-        vals = data[cid]
-
-        wmin = self.ui.value_min
-        wmax = self.ui.value_max
-
-        wmin.setText(pretty_number(nanmin(vals)))
-        wmax.setText(pretty_number(nanmax(vals)))
-
-    @property
-    def cmap(self):
-        combo = self.ui.color_scale
-        index = combo.currentIndex()
-        return combo.itemData(index)
+        self.ui.button_ok.clicked.connect(self.accept)
+        self.ui.button_cancel.clicked.connect(self.reject)
 
     def _apply(self):
+
         try:
-            lo, hi = self.vmin, self.vmax
+            lo, hi = self.state.v_min, self.state.v_max
         except ValueError:
             return  # limits not set. Abort
+
         if not np.isfinite(lo) or not np.isfinite(hi):
             return
 
-        subsets = facet_subsets(self._collect, self.component, lo=lo, hi=hi,
-                                steps=self.steps, log=self.log)
-        colorize_subsets(subsets, self.cmap)
+        subsets = facet_subsets(self._collect, self.state.att, lo=lo, hi=hi,
+                                steps=self.state.steps, log=self.state.log)
+        colorize_subsets(subsets, self.state.cmap)
 
     @classmethod
     def facet(cls, collect, default=None, parent=None):
-        """Class method to create facted subsets
-        The arguments are the same as __init__
+        """
+        Class method to create facted subsets.
+
+        The arguments are the same as __init__.
         """
         self = cls(collect, parent=parent, default=default)
         value = self.exec_()