+++ glue/dialogs/subset_facet/qt/tests/test_subset_facet.py
@@ -5,7 +5,7 @@ from matplotlib import cm
 
 from glue.core import Data, DataCollection
 
-from ..subset_facet import SubsetFacet
+from ..subset_facet import SubsetFacetDialog
 
 
 patched_facet = patch('glue.dialogs.subset_facet.qt.subset_facet.facet_subsets')
@@ -20,22 +20,22 @@ class TestSubsetFacet(object):
         self.s = dc.new_subset_group()
 
     def test_limits(self):
-        s = SubsetFacet(self.collect)
-        s.data = self.collect[0]
-        s.component = self.collect[0].id['x']
-        assert s.vmin == 1
-        assert s.vmax == 3
+        s = SubsetFacetDialog(self.collect)
+        s.state.data = self.collect[0]
+        s.state.att = self.collect[0].id['x']
+        assert s.state.v_min == 1
+        assert s.state.v_max == 3
 
     def test_get_set_cmap(self):
-        s = SubsetFacet(self.collect)
-        assert s.cmap is cm.cool
+        s = SubsetFacetDialog(self.collect)
+        assert s.state.cmap is cm.RdYlBu
 
     def test_apply(self):
         with patched_facet as p:
-            s = SubsetFacet(self.collect)
-            s.data = self.collect[0]
-            s.component = self.collect[0].id['x']
+            s = SubsetFacetDialog(self.collect)
+            s.state.data = self.collect[0]
+            s.state.att = self.collect[0].id['x']
             s._apply()
-            p.assert_called_once_with(self.collect, s.component,
+            p.assert_called_once_with(self.collect, s.state.att,
                                       lo=1, hi=3,
                                       steps=5, log=False)